Adding Gemfile.lock via bundle
From http://bundler.io/

Bundler maintains a consistent environment for ruby applications. It
tracks an application's code and the rubygems it needs to run, so that
an application will always have the exact gems (and versions) that it
needs to run.

aIf you were working on a gem, instead of an application, you would not
check-in your Gemfile.lock.

The following commands were run:
$ bundle
$ git add Gemfile.lock