Adding rspec-rails as a :test group dependency
Project Hydra traditionally makes use of the Rspec testing framework.

From http://rspec.info/
RSpec is testing tool for the Ruby programming language. Born under
the banner of Behaviour-Driven Development, it is designed to make
Test-Driven Development a productive and enjoyable experience with
features like:

* a rich command line program (the rspec command)
* textual descriptions of examples and groups (rspec-core)
* flexible and customizable reporting
* extensible expectation language (rspec-expectations)
* built-in mocking/stubbing framework (rspec-mocks)

By declaring rspec-rails in the :test group, the gem is only available
when the application is running in test.