Adding first spec step for viewing singup page
By using `rake routes` I was able to inspect what the existing routes
were for my Rails application. It looks like '/users/sign_in' is the
path to the Devise signup page.

Once the change was made, I ran `rspec` and an
ActiveRecord::StatementInvalid exception was raised (see below).

Rspec said that it could not find the 'users' table. This is because
our *test* database has not been created. When `rake db:migrate` is
run, only the *development* database was modified.

So I ran `rake db:test:prepare` and then ran `rspec`.

The following failure occurred when I ran `rspec`:

    Failures:

      1) user signup attempting to login without an account
         Failure/Error: visit('/users/sign_in')
         ActiveRecord::StatementInvalid:
           Could not find table 'users'

    Finished in 0.04759 seconds
    2 examples, 1 failure, 1 pending

    Failed examples:

    rspec ./spec/features/user_signup_spec.rb:4 # user signup…