Adding final step for validation
From https://github.com/plataformatec/devise#configuring-controllers

    Remember that Devise uses flash messages to let users know if sign
    in was successful or failed. Devise expects your application to
    call "flash[:notice]" and "flash[:alert]" as appropriate.

I was expecting my application to implement some sort of FlashHash
rendering. For more on Rails' FlashHash, see the following:
http://guides.rubyonrails.org/action_controller_overview.html#the-flash

Out of the box it does not. So I updated the application layout to
include a naive implementation of the flash handler. It was copied
directly from the Ruby on Rails Action Controller Overview Guide page.

But how did I know the warning message should be "Invalid email or
password." My original expectation was:

`page.should have_content('Missing')`

Once I added the FlashHash handling, the error message changed from:

    Capybara::ExpectationNotMet:
      expected to find css "div" with text "Missing." but there were
      no matches. Also found "", "Email", "Password", "Remember me",
      "", which matched the selector but not all filters.

To

    Capybara::ExpectationNotMet:
        expected to find css "div" with text "Missing." but there were
        no matches. Also found "Invalid email or password.", "",
        "Email", "Password", "Remember me", "", which matched the
        selector but not all filters.

Notice the addition of "Invalid email or password." That is the
message I was looking for.